requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
nltk>=3.8
spacy>=3.6.0
pandas>=1.5.0
requests>=2.31.0